emotion-music-app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
deepgram-sdk==2.11.0
fer==22.5.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.2.1
numpy==1.24.3
opencv-python==4.8.1.78